Filter Wizard: enforce limits on the baseband bandwidth
Now we coerce the baseband bandwith (BBBW) to be within the proper
ranges specified by the reference manual. If it's out of bounds, we
alter the calibration divider until the BBBW is back within the proper
range (or an error occurs if the divider value is skewed outside its
acceptable range).